Extract UpdatePos system
